rename utils to downloader
