+++ BiRefNet_node.py
@@ -17,7 +17,7 @@ from torchvision import transforms
 
 from loguru import logger
 from folder_paths import models_dir
-from utils import check_download_model
+from downloader import check_download_model
 
 config = Config()
