+++ models/backbones/build_backbone.py
@@ -5,7 +5,7 @@ from torchvision.models import vgg16, vgg16_bn, VGG16_Weights, VGG16_BN_Weights,
 from models.backbones.pvt_v2 import pvt_v2_b2, pvt_v2_b5
 from models.backbones.swin_v1 import swin_v1_t, swin_v1_s, swin_v1_b, swin_v1_l
 from config import Config
-from utils import check_download_model
+from downloader import check_download_model
 
 
 config = Config()
